+++ src/Provider/Eng/Color.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Fauxify\Fauxify\Provider\Eng;
+
+use Fauxify\Fauxify\Provider\ParentColor;
+
+class Color extends ParentColor{
+
+    protected $color = [
+        'Red', 'Green', 'Blue', 'Yellow', 'Black', 'White', 'Pink', 'Purple', 'Orange', 'Brown'
+    ];
+
+}