+++ src/Provider/Nep/Color.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Fauxify\Fauxify\Provider\Nep;
+
+use Fauxify\Fauxify\Provider\ParentColor;
+
+class Color extends ParentColor{
+
+    protected $color = [
+        'हरियो', 'निलो', 'पहेलो', 'कालो', 'सेतो', 'गुलाबी', 'बैगनी', 'सुन्तला', 'खैरो'
+    ];
+
+}